refactor merge files
